# swagger/swagger.json
{"swagger":"2.0","info":{"title":"Websocket test","description":"websocket試し","contact":{"name":"dfree1645","email":"d.free1645@gmail.com","url":"https://github.com/dfree1645"},"license":{"name":"MIT"},"version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/api/v1/ws":{"get":{"tags":["websocket"],"summary":"connect websocket","description":"websocket server","operationId":"websocket#connect","parameters":[{"name":"room","in":"query","description":"ルームID チャットルーム的な","required":true,"type":"string"}],"responses":{"101":{"description":"Switching Protocols"}},"schemes":["ws"]}},"/public/{filepath}":{"get":{"summary":"Download public/","operationId":"public#/public/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/swagger/{filepath}":{"get":{"summary":"Download public/swagger/","operationId":"swagger#/swagger/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}}},"definitions":{"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"SwitchingProtocols":{"description":"Switching Protocols"}},"externalDocs":{"description":"wiki","url":""}}